+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+* Set initial ruin threshold values to 10
+
 ## [0.0.16] - 2023-02-19
 
 ### Fixed